done component ButtonCart and added, fixed placholder in search
