+++ src/components/Button/Button.scss
@@ -4,7 +4,7 @@
 .Small__Btn {
   width: 32px;
   height: 32px;
-  margin: 0 12px;
+  margin-left: 12px;
   background-color: #fff;
   border: solid 1px $colorIcons;
   cursor: pointer;
