+++ src/components/ButtonFavor/ButtonFavor.scss
@@ -1,0 +1,23 @@
+@import "../../styles/reset";
+@import "../../styles/utils/mixins";
+@import "../../styles/utils/vars";
+
+.ButtonFavor {
+  height: 40px;
+  width: 40px;
+  background-color: #fff;
+  border: solid 1px $colorIcons;
+  @include iconInButton(url(../../img/Icons/heart.svg));
+  @include hover(border, 1px solid $colorPrimary);
+
+  &__Input {
+    opacity: 0;
+    width: 100%;
+    height: 100%;
+  }
+}
+
+.ButtonFavor--isInFavorites {
+  @extend .ButtonFavor;
+  @include iconInButton(url(../../img/Icons/redHeart.svg));
+}
