+++ src/components/ProductDetailsPage/ProductDetailsPage.tsx
@@ -9,6 +9,7 @@ import ProductsSlider from '../ProductsSlider/ProductsSlider';
 import { Breadcrumbs } from '../Breadcrumbs/Breadcrumbs';
 import { GoBack } from '../GoBack/GoBack';
 import ButtonFavor from '../ButtonFavor/ButtonFavor';
+import ButtonAddToCart from '../ButtonAddToCart/ButtonAddToCart';
 
 export const ProductDetailsPage: React.FC = () => {
   const [products, setProducts] = useState<Product[]>([]);
@@ -47,7 +48,8 @@ export const ProductDetailsPage: React.FC = () => {
 
   useEffect(() => {
     getProducts().then(data => setProducts(data));
-  }, []);
+    loadProductDetails(productId);
+  }, [productId]);
 
   useEffect(() => {
     if (products.length) {
@@ -56,10 +58,6 @@ export const ProductDetailsPage: React.FC = () => {
       setProductInfo(goodInfo);
     }
   }, [products, match.params.productId]);
-
-  useEffect(() => {
-    loadProductDetails(productId);
-  }, [productId]);
 
   const handleImages = (e: React.MouseEvent<HTMLElement>, i: number) => {
     e.preventDefault();
@@ -121,6 +119,7 @@ export const ProductDetailsPage: React.FC = () => {
     return value || 'unknown';
   };
 
+
   return (
     <>
       <section className="ProdactPage__Section">
@@ -131,8 +130,7 @@ export const ProductDetailsPage: React.FC = () => {
             Loading...
           </div>
         )}
-        {isLoading && isLoaded && ''}
-        {productDetails && (
+        {isLoaded && productDetails && productInfo && (
           <article className="ProdactPage">
             <Breadcrumbs />
             <GoBack />
@@ -184,18 +182,11 @@ export const ProductDetailsPage: React.FC = () => {
                     )}
                   </div>
                   <div className="Card__ButtonWrap">
-                    <button
-                      type="button"
-                      className="Card__ButtonCart--width"
-                    >
-                      Add to cart
-                    </button>
-                    {/* <button
-                      type="button"
-                      className="Card__ButtonFavor"
-                      aria-label="Mute text"
-                    /> */}
-                    <ButtonFavor product={productInfo}/>
+                    <ButtonAddToCart
+                      product={productInfo}
+                      ClassNameForBtn="ButtonCart--width"
+                    />
+                    <ButtonFavor product={productInfo} />
                   </div>
                 </section>
                 <section className="ProdactPage__Info">
